+++ NEWS.md
@@ -6,6 +6,14 @@
 * In `plot_confusion_matrix()`, adds argument `class_order` for manually setting the order of the classes
 in the facets.
 
+* In `plot_confusion_matrix()`, tiles with a count of `0` no longer has text in the tile by default.
+This adds the `rm_zero_percentages` (for column/row percentage) and `rm_zero_text` (for counts and normalized) arguments. 
+
+* In `plot_confusion_matrix()`, adds optional sum tiles. Enabling this (`add_sums = TRUE`) adds an extra column and
+an extra row with the sums. The corner tile contains the total count. This adds the `add_sums` and `sums_settings` arguments. A `sum_tile_settings()` function has been added to control the appearance of these tiles.
+
+* In `plot_confusion_matrix()`, adds option (`intensity_by`) to set the color intensity of the tiles to the overall percentages (`normalized`). 
+
 # cvms 1.1.0
 
 * In `plot_confusion_matrix()`, adds option to only have row and column percentages in the diagonal tiles. Thanks to [@xgirouxb ](https://github.com/LudvigOlsen/cvms/issues/12) for the idea.